Refactor experiments folder structure
