+++ experiments/synthetic_eval.sh
@@ -1,0 +1,12 @@
+SYN_DIR=synthetic_exp
+
+for dir in $(ls $SYN_DIR/data)
+do
+	mkdir -p $SYN_DIR/results/$dir
+	echo Evaluating $dir
+	for mode in SAPA PA XSDD XADD FXSDD
+	do
+		echo Mode $mode
+		python3 evaluateModels.py $SYN_DIR/data/$dir -o $SYN_DIR/results/$dir -m $mode 
+	done
+done
